worker/requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==1.1.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
psycopg2-binary==2.9.4
flask_cors==1.9
flask_jwt_extended==4.0.0
celery==4.4.3
gunicorn==20.1.0
redis==4.3.4
pydub==0.25.1
google-cloud-storage==2.5.0
google-cloud-pubsub== 2.13.11
